refactor: reorganize imports and clean up unused references
