+++ src/koheesio/spark/transformations/__init__.py
@@ -337,7 +337,6 @@ class ColumnsTransformation(Transformation, ABC):
         df = df or self.df
         if not df:
             raise RuntimeError("No valid Dataframe was passed")
-        from koheesio.spark.utils.connect import Column
 
         if not isinstance(col, Column):
             col = f.col(col)
