+++ tests/spark/readers/test_metastore_reader.py
@@ -1,5 +1,6 @@
 import pytest
 
+from koheesio.spark import DataFrame
 from koheesio.spark.readers.metastore import MetastoreReader
 
 pytestmark = pytest.mark.spark
@@ -9,8 +10,6 @@ def test_metastore_reader(spark):
     df = MetastoreReader(table="klettern.delta_test_table").read()
     actual = df.head().asDict()
     expected = {"id": 0}
-
-    from koheesio.spark.utils.connect import DataFrame
 
     assert isinstance(df, DataFrame)
     assert actual == expected
